--- client/components/AddListing copy.jsx
@@ -1,10 +1,0 @@
-import React from 'react';
-
-const AddListing = () => (
-  <div>
-    <h2>This is the Add Listing Component></h2>
-  </div>
-);
-
-export default AddListing;
-
