+++ client/components/listings.jsx
@@ -1,24 +1,9 @@
-import ListingEntry from './listingEntry.jsx';
-import ListingInfo from './listingInfo.jsx';
+import React from 'react';
 
-const Listings = props => {
-  let filtered = [];
-  //Iterate over all listings and generate filtered population based on activeFilter value
-  props.listings.forEach(listing => {
-    if (props.activeFilter === 'All' || listing.location === props.activeFilter) {
-      filtered.push(listing);
-    }
-  });
-  //Return the filtered population only & show a specific listing's info if it has been clicked
-  return (
-    <div className="listings">
-      {filtered.map(listing =>
-        (props.activeListing && props.activeListing === listing.listingId) ?
-        <ListingInfo handleListingInfoClick = {props.handleListingInfoClick} listing={listing} user={props.user}/> :
-        <ListingEntry key={listing.listingId} handleListingEntryClick={props.handleListingEntryClick} listing={listing}/>)
-      }
-    </div>
-  );
-};
+const Listings = () => (
+  <div>
+    <h2>This is the Add Listing Component></h2>
+  </div>
+);
 
 export default Listings;
